Validate comment

+++ app/Http/Controllers/EmployeeController.php
@@ -10,52 +10,69 @@ use Illuminate\Support\Facades\DB;
 use App\Employee;
 use App\Emp_dep_position;
 
+//Excel Imports & Exports Declaration
 use Maatwebsite\Excel\Facades\Excel;
 // use App\Imports\EmployeesImport;
 use App\Exports\EmployeesExport;
 
-//mail**
+
+//mail declaration** 
 use Illuminate\Support\Facades\Mail;
 
 //paginate assign
 use Illuminate\Support\Facades\Config;
 
+//validation
+use App\Http\Requests\StoreMarket;
+use Illuminate\Routing\Controller;
+
 
 class EmployeeController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Display a listing of the employee with positon id and departments id.
+     * Author
+     * 26/8/2020
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        //
-        // $employees=Employee::all();
+
         $perPage=Config::get('constant.per_page');
-
-        $limit=(int)env('limit');
-        // $employees=Employee::paginate($limit);
-        $employees=Employee::with('department','position')->onlyTrashed()->paginate($perPage);
-
+        
+        $employees=Employee::with('department','position')->withTrashed()->paginate($perPage);
         
 
-        $employees=Employee::with('department','position')->onlyTrashed()->get();
-        return ['employees'=>$employees,];
+        // $employees=Employee::with('department','position')->onlyTrashed()->first();
+        return ['employees'=>$employees];
         
     }
 
     /**
      * Store a newly created resource in storage.
+     * 
+     * 26/8/2020
      *
      * @param  \Illuminate\Http\Request  $request
+     * @param employee_name,email,dob,password,gender,position_id,department_id
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         //
         try{
-            
+            $data = $request->validate([
+                'employee_name' => 'bail|required|max:20|alpha',
+                'email' => 'bail|required|unique:employees,email',
+                'dob' =>'bail|nullable|date:Y-M-D|before:18 years ago',
+                'password' =>'required|min:8',
+                'gender' =>'required',
+                'position_id' =>'required',
+                'department_id' =>'required',
+                
+            ]);
+
             $employees = new Employee();
             $employees->employee_name = $request->employee_name;
             $employees->email = $request->email;
@@ -63,19 +80,21 @@ class EmployeeController extends Controller
             $employees->password = Hash::make($request->password);
             $employees->gender = $request->gender;
             $employees->save();
-            // 
+            // after employee save ,get maximum employee id
             $lastemp_id = Employee::max('id');
+            // Position id is exist
             if($request->position_id){
                 $position_id=$request->position_id;
 
             }else{
-                    $position_id=1;
+                    $position_id=Config::get('constants.position_id');
                 }
+                // Department id is exist
                 if($request->department_id){
                     $department_id=$request->department_id;
 
                 }else{
-                        $department_id=1;   
+                        $department_id=Config::get('constants.department_id'); 
                     }
                     
                     
@@ -106,6 +125,8 @@ class EmployeeController extends Controller
 
     /**
      * Display the specified resource.
+     * 
+     * 27/8/2020
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
@@ -122,6 +143,8 @@ class EmployeeController extends Controller
 
     /**
      * Update the specified resource in storage.
+     * 
+     * 26/8/2020
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -132,6 +155,15 @@ class EmployeeController extends Controller
         //
         try{
 
+            $data = $request->validate([
+                'employee_name' => 'bail|required|max:20',
+                'email' => 'bail|required|unique:employees,id',
+                'dob' =>'required',
+                'password' =>'required',
+                'gender' =>'required',
+                
+            ]);
+
             $employees = Employee::find($id);
 
             $employees->employee_name = $request->employee_name;
@@ -139,7 +171,7 @@ class EmployeeController extends Controller
             $employees->dob = $request->dob;
             $employees->password = $request->password;
             $employees->gender = $request->gender;
-            $employees->save();
+            $employees->update();
 
             // $lastemp_id = Employee::max('id');
             $emp=Emp_dep_position::where('employee_id',$id)->first();
@@ -149,24 +181,34 @@ class EmployeeController extends Controller
                     $position_id=$request->position_id;
 
                 }else{
-                        $position_id=1;
+
+                    return response()->json([
+                        'message' =>'Data not found Position Id'],200);
+                        // $position_id=1;
                     }
 
                     if($request->department_id){
                         $department_id=$request->department_id;
 
                     }else{
-                            $department_id=1;   
+                        return response()->json([
+                            'message' =>'Data not found Position Id'],200);
+                            // $department_id=1;   
                         }
 
                        
                 $emp->department_id =$department_id;
                 $emp->position_id =$position_id;
                 $emp->update();
+
+                return response()->json([
+                    'message' =>'Successful'],200);
+            }else{
+                return response()->json([
+                    'message' =>'Data not found'],200);
             }
 
-            return response()->json([
-                'message' =>'Successful'],200);
+           
             
         }catch(QueryException $e){
             return response()->json([
@@ -176,8 +218,10 @@ class EmployeeController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
+     * Soft-Delete from Employee and Emp_dep_position tables
+     * 
+     * 27/8/2020
+     * 
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
@@ -188,15 +232,59 @@ class EmployeeController extends Controller
         $emp_dep_position=Emp_dep_position::where('employee_id',$id)->FirstOrFail()->delete();
         
     }
+
+    /**
+     * Force-Delete from Employee and Emp_dep_position tables
+     * 
+     * 27/8/2020
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
     public function forceDelete($id)
     {
         //
-    
-        $employees = Employee::withTrashed()->where('id', $id)->forceDelete();
-        $emp_dep_position = Emp_dep_position::withTrashed()->where('employee_id', $id)->forceDelete();
-        return false;
+
+        try{
+            $employees = Employee::withTrashed()->where('id', $id)->FirstOrFail();
+            if($employees){
+                $employees = Employee::withTrashed()->where('id', $id)->forceDelete();
+            }else{
+                return response()->json([
+                    'message' =>'Employee Id not found'],200);
+            }
+
+            $emp_dep_position = Emp_dep_position::withTrashed()->where('employee_id', $id)->FirstOrFail();
+            if($emp_dep_position){
+            $emp_dep_position = Emp_dep_position::withTrashed()->where('employee_id', $id)->forceDelete();
+            
+            }else{
+                return response()->json([
+                    'message' =>'Employee Id not found in emp_dep_position'],200);
+            }
+            return response()->json([
+            'message' =>'Successful'],200);
+        
+        }catch(Exception $e){
+            return response()->json([
+                'message' =>$e->getMessage()
+                
+            ]);
+
+            echo "Message: " . $e->getMessage();
+        }
        
     }
+
+    /**
+     * Seacrch Employee with Departments and Positions Tables
+     * 
+     * 27/8/2020
+     * 
+     * @param  \Illuminate\Http\Request  $request-> id, employee_name
+     * 
+     * @return \Illuminate\Http\Response
+     */
 
     public function search(Request $request){
 
@@ -221,6 +309,15 @@ class EmployeeController extends Controller
 
         
     }
+    /**
+     * Excel File Export 
+     * 
+     * 28/8/2020
+     * 
+     * @param  \Illuminate\Http\Request  $request-> id, employee_name
+     * 
+     * @return \Illuminate\Http\Response
+     */
     public function fileExport(Request $request)
     {
     
@@ -239,5 +336,61 @@ class EmployeeController extends Controller
         return Excel::download(new EmployeesExport($search_data), 'MttRegistrations.xlsx');
         
     }
+
+    public function excel(Request $request)
+	{
+        $search_data=[];
+        if($request->id){
+            $search_id=['id',$request->id];
+            array_push($search_data,$search_id);
+
+        }if($request->name){
+
+                $search_name=['employee_name','LIKE',$request->name.'%'];
+                array_push($search_data,$search_name);
+
+            }
+
+        $limit=(int)env('limit');
+        $employees=Employee::with(['department','position'])
+                            ->withTrashed()
+                            ->where($search_data)
+                            ->first();
+                            
+                        
+        
+
+        // dd($employees);
+		// $formal_data = Employee::table('formal')->get()->toArray();
+		$programma_array[] = array('employee_name', 'email', 'dob', 'password', 'Gender');
+			// dd($employees);
+		// foreach($employees as $pass)
+		// {
+            
+			$programma_array[] = array
+			(
+			'employee_name'=> $employees->employee_name,
+			'email'=> $employees->email,
+            'dob'=> $employees->dob,
+            'password'=> $employees->dpassword,
+			'Gender'=> $employees->gender
+      		);
+			
+		//}
+		// dd($pass);
+			
+			Excel::create('Programma_Data', function($excel) use ($programma_array)
+			{
+				$excel->setTitle('Apotelesmata');
+				$excel->sheet('Programma_Data', function($sheet) use ($programma_array)
+				{
+					$sheet->fromArray($programma_array, null, 'A1', false, false);
+				});
+			}) -> download('xlsx');
+			
+	        
+    }
+			   
 }
+
